getProperty move to native
