+++ trustdevice/src/main/cpp/detection_risk.h
@@ -15,8 +15,8 @@ typedef unsigned long operation_type;
 static unsigned long trampoline_code = 0x58000050;
 #endif
 
-extern "C" JNIEXPORT jint JNICALL detect_debug(JNIEnv *env, jclass clazz);
-extern "C" JNIEXPORT jstring JNICALL detect_hook(JNIEnv *env, jclass clazz);
+extern "C" JNIEXPORT jint JNICALL detect_debug(JNIEnv *env, jobject clazz);
+extern "C" JNIEXPORT jstring JNICALL detect_hook(JNIEnv *env, jobject clazz);
 
 class detection_risk {
