+++ trustdevice/src/main/cpp/utils.h
@@ -6,6 +6,7 @@
 #define TRUSTDEVICE_ANDROID_UTILS_H
 
 #include <android/log.h>
+
 #define TAG "TrustDevice"
 
 #define LOGI(...)  __android_log_print(ANDROID_LOG_INFO, TAG, __VA_ARGS__)
@@ -13,8 +14,7 @@
 #define LOGD(...)  __android_log_print(ANDROID_LOG_DEBUG, TAG, __VA_ARGS__)
 #define LOGW(...)  __android_log_print(ANDROID_LOG_WARN, TAG, __VA_ARGS__)
 
-class utils {
-};
-
+extern "C" JNIEXPORT jstring JNICALL
+get_property(JNIEnv *env, jobject clazz, jstring _key, jstring _default_value);
 
 #endif //TRUSTDEVICE_ANDROID_UTILS_H
