+++ trustdevice/src/main/java/cn/tongdun/mobrisk/core/tools/JNIHelper.kt
@@ -8,4 +8,5 @@ package cn.tongdun.mobrisk.core.tools
 object JNIHelper {
     external fun detectDebug(): Int
     external fun detectHook(): String
+    external fun getProperty(key: String, defaultValue: String): String
 }
